Restructure lesson amd add minor changes to tasks 24-27
