--- src/lesson_5/Task26/classes/Refrigerator.java
@@ -1,10 +1,0 @@
-package lesson_5.Task26.classes;
-
-import lesson_5.Task26.abstracts.Refrigerators;
-
-public class Refrigerator extends Refrigerators {
-    public Refrigerator(String name) {
-        super(name);
-    }
-
-}
