--- src/lesson_5/Task26/interfaces/IPLug.java
@@ -1,12 +1,0 @@
-package lesson_5.Task26.interfaces;
-
-public interface IPLug {
-
-
-    void plugInMains();
-
-    void unplugFromMains();
-
-    void setTemperature();
-
-}
