--- src/lesson_5/Task26/interfaces/IStarter.java
@@ -1,9 +1,0 @@
-package lesson_5.Task26.interfaces;
-
-public interface IStarter {
-
-    void pressStart();
-
-    void pressStop();
-
-}
