--- src/lesson_5/Task27/abstracts/BankCard.java
@@ -1,49 +1,0 @@
-package lesson_5.Task27.abstracts;
-
-
-import lesson_5.Task27.enums.CardBrand;
-import lesson_5.Task27.enums.CardType;
-import lesson_5.Task27.enums.Currency;
-
-import java.text.DecimalFormat;
-import java.text.DecimalFormatSymbols;
-import java.util.Locale;
-
-public abstract class BankCard {
-
-    protected final CardBrand CARD_BRAND;
-    protected final String CARD_NUMBER;
-    protected final Currency CURRENCY;
-    protected final Date DATE;
-    protected double cardAmount;
-    protected CardType CARD_TYPE;
-
-    protected DecimalFormatSymbols symbols = new DecimalFormatSymbols(Locale.US);
-    protected DecimalFormat df = new DecimalFormat("##.##", symbols);
-
-
-    public BankCard(CardBrand CARD_BRAND, String CARD_NUMBER, Currency CURRENCY, Date DATE, double cardAmount) {
-        this.CARD_BRAND = CARD_BRAND;
-        this.CARD_NUMBER = CARD_NUMBER;
-        this.CURRENCY = CURRENCY;
-        this.DATE = DATE;
-        this.cardAmount = cardAmount;
-    }
-
-
-    public void depositCash(double amount) {
-        System.out.println(df.format(amount) + " has been added to your card account");
-        cardAmount += amount;
-    }
-
-    public void withdrawCash(double amount) {
-        System.out.println(df.format(amount) + " has been withdrawn from your card account");
-        cardAmount -= amount;
-    }
-
-    @Override
-    public String toString() {
-        return "BankCard [Type: " + CARD_TYPE + ", Name: " + CARD_BRAND + ", Card Number: " + CARD_NUMBER
-                + ", Currency: " + CURRENCY + ", Expiration Date: " + DATE + ", Amount: " + df.format(cardAmount) + ']';
-    }
-}
