--- src/lesson_5/Task27/enums/CardBrand.java
@@ -1,6 +1,0 @@
-package lesson_5.Task27.enums;
-
-public enum CardBrand {
-    MASTERCARD, VISA_CLASSIC, VISA_GOLD;
-
-}
