--- src/lesson_5/Task27/enums/CardType.java
@@ -1,7 +1,0 @@
-package lesson_5.Task27.enums;
-
-public enum CardType {
-
-    GIFT_CARD(), DEBIT_CARD (), CREDIT_CARD ();
-
-}
