--- src/lesson_5/Task27/enums/Currency.java
@@ -1,7 +1,0 @@
-package lesson_5.Task27.enums;
-
-public enum Currency {
-
-    USD, BYN, EUR
-
-}
