+++ src/lesson_5/com/maintasks/Task25/CashMachineRunner.java
@@ -1,0 +1,18 @@
+package lesson_5.com.maintasks.Task25;
+
+/*
+Создать класс и объекты описывающие Банкомат. Набор купюр находящихся в банкомате должен задаваться тремя свойствами:
+количеством купюр номиналом 20 50 100. Сделать методы для добавления денег в банкомат. Сделать функцию снимающую деньги.
+На вход передается сумма денег. На выход – булевское значение (операция удалась или нет).
+При снятии денег функция должна рапечатывать каким количеством купюр какого номинала выдается сумма.
+Создать конструктор с тремя параметрами – количеством купюр.
+*/
+
+public class CashMachineRunner {
+    public static void main(String[] args) {
+
+        CashMachine cashMachine = new CashMachine(300, 150, 50);
+        cashMachine.displayMenu();
+
+    }
+}
