+++ src/lesson_5/com/maintasks/Task26/classes/Refrigerator.java
@@ -1,0 +1,10 @@
+package lesson_5.com.maintasks.Task26.classes;
+
+import lesson_5.com.maintasks.Task26.abstracts.Refrigerators;
+
+public class Refrigerator extends Refrigerators {
+    public Refrigerator(String name) {
+        super(name);
+    }
+
+}
