+++ src/lesson_5/com/maintasks/Task26/interfaces/IPLug.java
@@ -1,0 +1,12 @@
+package lesson_5.com.maintasks.Task26.interfaces;
+
+public interface IPLug {
+
+
+    void plugInMains();
+
+    void unplugFromMains();
+
+    void setTemperature();
+
+}
