+++ src/lesson_5/com/maintasks/Task26/interfaces/IStarter.java
@@ -1,0 +1,9 @@
+package lesson_5.com.maintasks.Task26.interfaces;
+
+public interface IStarter {
+
+    void pressStart();
+
+    void pressStop();
+
+}
