+++ src/lesson_5/com/maintasks/Task27/enums/CardBrand.java
@@ -1,0 +1,6 @@
+package lesson_5.com.maintasks.Task27.enums;
+
+public enum CardBrand {
+    MASTERCARD, VISA_CLASSIC, VISA_GOLD;
+
+}
