+++ src/lesson_5/com/maintasks/Task27/enums/CardType.java
@@ -1,0 +1,7 @@
+package lesson_5.com.maintasks.Task27.enums;
+
+public enum CardType {
+
+    GIFT_CARD(), DEBIT_CARD (), CREDIT_CARD ();
+
+}
