+++ src/lesson_5/com/maintasks/Task27/enums/Currency.java
@@ -1,0 +1,7 @@
+package lesson_5.com.maintasks.Task27.enums;
+
+public enum Currency {
+
+    USD, BYN, EUR
+
+}
